Migrate fjs/djs/transpiler to .f.mjs
Implementation and proof migrate from authored TypeScript to
JSDoc-typed JavaScript under the stage-1 migration. ParseContext and
djsResult stay as JSDoc typedefs in the module (their original public
export status preserved) — nothing outside the module imports them by
name today. transpile's step callback and proof's run helper get
explicit return-type annotations, matching the original's, since
TypeScript can't otherwise reconcile the union across their two
return branches. Updates fjs/djs/module.f.ts's import specifier — its
own runtime dependencies (serializer, transpiler) are now both
migrated — and fixes a stale module.f.ts doc reference in
fjs/effects/todo/effect-list-fold.md.
